Add Tunnel Zone and Tunnel Zone Host API handlers
This patch adds TunnelzoneHandlerMixin and
TunnelzonehostHandlerMixin to the core plugin using
generate_methods decorators.

This patch also adds symbols for 'list', 'show', 'create', 'update'
and 'delete'. It'd be too easy to have typos for them and I added
symbols for them to make it easier to detect typos using the static
checker.

Change-Id: Ia652c5afa0e00a358325b80d6c6dab27152ced56
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>