Modify generate_methods to accept predefined alias
This patch modifies generate_methods to accept predefined
aliases of the API handlers. Previously it uses the first Camel
or Camel_snake cases in the class name and it does now, but it
also supports the predefined ALIAS attribute in the class
definition and use it to generate methods over its class name.

Change-Id: I1480d839a68359c84704ab7bdfcb0c3502b70f0b
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>