Add API handlers for system state
This patch adds the API handlers for the system state of
MidoNet.

The current system state follows the rule which defines the URI
structure in the strict REST style because of the constraint of
Neutron's framework. So we need to call the appropriate
endpoints such as "/systems/midonet" rather than "/sytem_state",
which was the original URI in the Java implementation.

Change-Id: I4deaa71a50883d7d0f9d2d9db84e15c682303a33
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>