Fix a critical bug of delete_port
This patch fixes a bug of delet_port which is calling
disassocate_floatingips with a keyword argument do_notify=False of the
parent class, L3_NAT_db_mixin. It caused TypeError and kills the
plugin in the old version, where the keyword argument is not added.

In the previous version, before 2014.1.2 specifically, L3_NAT_db's
delete_port doesn't take any keyword argument. So I replace this call
with another one eliminating do_notify=False. In the current version it
defaults to True if it's not specified.

This fixes MN-3010.

Change-Id: Ie472e727d4bdcbf031edf058cbae1441447fce75
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>