Fixing a regex bug in version_tag
version tag for unstable builds assumed that the last tag returned by:
`get describe --tags` always returned a {version}.rcX tag value. This
can not be true for maintain versions, where the version released is not
'rc' anymore. Change the regex to make the 'rc' part optional.

Furthermore, if the commit is not the tag by itself, will print out the
tag name, a hyphen, the number of commits made, a hyphen, the letter ‘g’
and then the commit identifier. This way:

    `2014.2+1.0-6-gc1433d2`

add the `6-gc1433d2` part to the regex.

Change-Id: I1f6fae6a00e0d5aef7a0d42ada70145d34bcf4e3