Add 'testr' command via 'pretty_tox' in tox
Adding the execution of unit tests when running `tox -v -epy27` in
gates. It was deactivated before. `pretty_tox` script is copied from
current master neutron. `pretty_tox` uses `subunit-trace` which is part
of the `tempest_lib` dependency. This dependency it allow us to open the
door to future functional tests in the plugin.

Change-Id: I066131060ec37dd5513f1ca5b2493517ea92ea8b