Add provider extension support
Adds support for the 'provider:network_type' field in network
creation. This is included in the 'provider' extension.

Other fields included in this extension (segmentation_id and physical_network)
are not supported.

The only value allowed for 'provider:network_type' is 'local'.

This means that the user is allowed to specify that a network with the
fields 'provider:network_type':'local' if he chooses. This information
is then given to the midonet backend so that the backend can treat
this network as a 'provider network', which is described in specs/provider_net.

Co-Authored-By: Joe Mills <joe@midokura.com>

Implements: specs/provider_net
Change-Id: I475388612a941b6511da1781bca2971270052abb
Fixes: MNP-73