Avoid pbr to modify version to 'dev'
`pbr` dependency adds a lot of cool things when it comes to create
source distributions: it adds ChangeLog, AUTHORS, sphinx autodoc...

However, it also creates a too-much strict python naming version value,
by adding 'postX' or 'devY' to the egg version. This is annoying if we
want to offer rpms and debs.

Adding PKG-INFO file in the root's project source code, avoids this
behaviour.

Change-Id: If7940395b3f99cbd37195ef93d4735c007ea6680