Replace MIDONET_ENABLE_Q_SVC_ONLY with sample conf
MIDONET_ENABLE_Q_SVC_ONLY env variable is used as a convenient way to
only start Neutron services. However, the same could be achieved simply
by specifying in local.conf which services to run, and by doing so, it
would be more explicit to the users which ones are actually running.
This patch replaces MIDONET_ENABLE_Q_SVC_ONLY with a sample local.conf
file that does the following:

 * Start the core OpenStack dependencies (mysql, rabbit, keystone)
 * Start Neutron, Nova, Glance
 * Start Tempest (since this would be useful in CI and local dev)

Change-Id: I4f574fe2c543d5016edaa577c6b17f38313c3971
Implements: MNP-128