Set the supported extensions in __init__
Instead of setting the supported extension aliases when the plugin class
is loaded, do so when the plugin object is created.

Without this patch, the unit tests for extra extensions required
unnecessary extra care with regards to the timing of the plugin module
import and the configuration overrides.

More concretely, in the unit tests, the configuration overrides happen
in the test setup, and that's when the extra extensions get registered.
However, since the Neutron plugin class accesses the
supported_extension_aliases of the MidoNet mixin class statically, as
soon as the plugin module loads (i.e. imported), the supported aliases
get set before the configurations could be modified.

Change-Id: I455474dd89c1a42805a5b6b68a4d5514161ad40f
Fixes: MNP-135