Split Midonet Plugin into legacy and new arch
Kilo plugin is compatible with both the legacy ZK-based MidoNet and the
new ZOOM-based MidoNet. Since they differ in the list of features,
split them into two plugins, v1 (legacy) and v2 (ZOOM).

The v1 and v2 plugins are completely separated and only share the
initialization code since v1 is considered to be feature frozen, and
they should evolve independently.

Even though v1 does not really need the pluggable client model, it is
kept this way so that it is easier to integrate with LBaaS service
which uses the same pluggable client model.

Change-Id: I5fda4cb5af2a1097efb3362b2920a64a538e7623
Implementes: MNP-143