Use UPLINK net type instead of LOCAL
Previously, 'LOCAL' network type was used to determine whether a
Neutron network is an uplink network or not. This caused a problem in
Neutron because, following the behavior of ML2, all external networks
when 'providernet' extension is enabled usually have a type, and
typically configured to have 'LOCAL' as the default, unless explicitly
specified, making it impossible to differentiate uplink and external
networks.  To get around it, this patch defines a new type 'UPLINK' to
replace 'LOCAL'.  Also, instead of validating against the unsupported
types, just log and continue.  This way, 'providernet' extension could
continue to be used with Horizon, until it is patched to accept uplink
type.

Change-Id: Ifb0dbe8d2c86456b11064c786f8ebe6e4126d206
Implements: MNP-150