Validate subnet for pool/vip properly
There was a validation to prevent VIP and Pool that belong in the same
subnet to be created. This was wrong because LB would work even if they
are in the same subnet as long as the client is not in the same subnet
as the VIP.  The validation should have occurred only when there as a
health monitor associated with a pool  This patch implementes the
correct validation.

Change-Id: Ib2dd24048a6badedb8baf7b3625d47f0e87f34b2
Fixes: MNP-154