Fix packaging issue by removing ${python:Depends}
Debian `control` file establishes the dependencies of the package. One
function that can be used is the ${python:Depends}, which will create
`python-*` dependencies based on `requirements.txt`. So it was defining
`python-oslo.log` as dependency automatically.

Problem is `python-oslo.log` does not exist as a package in Debian,
correct name is `python-oslo-log`. So better to remove the
${python:Depends} and define all the dependencies explicitly.

Change-Id: Idab8899e5fe47c1c9b7397ede68c2dbb2b7c884d