PASS 2 unsupported ipv6 UT tests
These tests are expected to fail due to how we handle transactions.
pass these tests until we have fixed the transaction problem.

Another change, but unrelated to the first is updating the
requirements.txt file. Without explicitly setting the oslo.serialization
requirements, version 1.7.0 will be downloaded and used for UT. However,
stable/kilo does not support this version of oslo.

Both of these changes are required to pass UT on stable/kilo.

The tests that need to be skipped are already skipped in master
as a part of a different, but unrelated patch with commit id:

09a0a309a2c94cc559af24d317e94ede1d863860

summary of changes:
devstack/settings: change default branch of midonet to master.
test_midonet_plugin.py: pass the 2 tests that are not supported currently.
test_midonet_plugin_v2.py: pass the 2 tests that are not supported currently
    for version 2 of the plugin.
requirements.txt: Update the requirements to be in sync with stable/kilo
    requirements.

Change-Id: I41a99882db5259a40f56cf3b3bffa584ca21d154
Closes-Bug #1480378
Signed-off-by: Joe Mills <joe@midokura.com>
Co-Authored-By: Ryu Ishimoto <ryu@midokura.com>
Co-Authored-By: YAMAMOTO Takashi <yamamoto@midokura.com>