Add LBaaS unit tests and validations
Unit tests never existed for LBaaS in the Juno plugin, so adding them
in this patch.  This also exposed key validations missing in the
plugin, so they were also added.

This patch is going directly to the stable branch for the following
reasons:

 * All the validations added in this patch already exist in master
 * LBaaS implementation changed significantly between Juno and Kilo for
the MidoNet plugin so it is impossible to cherry-pick.

Change-Id: If2c83b69f1636e0f726fbe4a1f2dbf19216f027a
Fixes: MNP-179