Add references to SG callbacks to avoid GC
ML2 mechanism driver for MidoNet is currently used in both the Liberty
and Kilo environment, and must continue to support both cases.

In Kilo, weakref was used for storing callback methods where they would
get garbage collected, making all callbacks registered unavailable.
This problem was fixed in Liberty with the commit 8959032.

Until post-Liberty when we no longer need to support the mech driver
compatibility with Kilo ML2, add explicit references to the callback
methods so that they are not garbage collected.

Revert this patch when there is no need to support compatibility with
Kilo ML2.

Change-Id: I71b34b61624fbb40e1338eb6b267744e5d1e85fb
Closes-Bug: #1504373