Rely on pbr to define the version of the library
Pbr has a way to create the library version based on last tag. If
current commit does not match with the tag, it adds the PEP 440
compliance value devXXX where XXX is the number of commits after the
last tag, highlighting that it is a development version.

Before 2015.1, MidoNet had tags not recognized by pbr. However, adding
the PKG-INFO file we could trick pbr to set the version number we
wanted.

Now we are pbr and PEP440 compliance, so we can rely on 'pbr' to create
this version. It is time to improve setup.cfg and delete the PKG-INFO
file, which will be generated by the `pbr` tool.

Change-Id: Id3d83611ba2d87467a5d522891b6dd5f83e8f9a6