plugin_v2: Deal with the lack of portsecurity DB rows
For get methods, consider the lack of the binding row as
the API default value of network portsecurity.
(It's similar to ML2.  The reason it was necessary
for ML2 is a bit different from midonet, though.)

For update methods, create the binding row for the resource
if it isn't found.

Similar ML2 change: I73abc98d83372f6259f17680806e6541458e2077
Closes-Bug: #1507651
Change-Id: I97b128f2801b7bc5bf1ff028b7927bda068643d8