Re-enable and improve VPNaaS tests
There were VPNaaS tests that were disabled because of an upstream bug.
It's already solved so re-enable them again. Besides, this patch makes
the vpn service and ipsec site connection update tests more specific
to avoid regressions, as we also check that the dtos sent to the midonet
backend actually contain the fields updated.

Change-Id: I771730f6f356c55bf16935548b3fc107d8f38cad
Signed-off-by: Xavi León <xavi.leon@midokura.com>