l3: Remove device_owner validation for FIP association
The purpose of the check was to avoid user confusion due to
neutron db vs backend mismatch.  Given that we use the ERROR state [1]
these days, it's no longer too useful.

On the other hand, the recent versions of kuryr use a device_owner
which is not known to networking-midonet or midonet. ("kuryr:container")
This commit removes the check to allow such a usage.

Note: As far as the cluster rest api is used, MidoNet handles unknown
device_owner values as COMPUTE.  (Task-based api is a different story
though)

[1] Id3bfb87c80f66293340bee0eff5c77eb64dbd522

Closes-Bug: #1610488
Related-Bug: #1535211
Change-Id: I873ea39c029cded09a5c24e497846ec62d292638