FIP6: Support creating IPv6 FIP
This modifies l3_db_midonet to override the _create_floatingip
method in order to allow the creation of floating IPs with IPv6
addresses on an external network. If the external network has
dual-stack IPv4/IPv6 subnets, creating the floating IP for the
IPv4 subnet takes precedence over the IPv6 subnets.

Related-Bug: #1614477
Co-Authored-By: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Ia852002aaad32e5c91c45ecf5f93a8c0c432c1b7