Cleanup tox_install_project artefacts
Translation sync often fails for networking-midonet since
/tmp/openstack/neutron already exists.

This happens since we run the translation sync on a machine that is used
for a variety of jobs, a previous run of this job will leave the content
there and then fail.

Install inside the source directory, so that git clean will work.

Note this is a partial sync with the same file in networking-odl.

Change-Id: Ie6e72eab9836f2c2fe4e56a65f97abd477417115
Closes-Bug: #1656506