Disable custom plugin scenarios in rally
The rally jobs are failing because they fail to validate
the subtasks that use our custom plugins. It seems that
something has changed with rally upstream and we need to
fix our plugins in order to met the new requirements.

This patchs remove the scenarios that are using our custom
plugins.

Related-Bug: #1670577

Change-Id: I7d8430dc988a68bc6f465f7e6035237da5fb261b
Signed-off-by: Antonio Ojea <aojea@midokura.com>