v2: Replace security group implementation with ML2 version
The recent Neutron change [1] introduced a race in our
v2 plugin for the default security group creation.
Instead of trying to adapt the v2 plugin code,
just switch the implementation to the one in our ML2 driver.

Note: This breaks the task-based client because the
implementation doesn't call the client precommit methods.

[1] Iaa83c8664d5bfde31fdcdd694f6f18d9ef9bf14a

Closes-Bug: #1671848
Change-Id: I17262ecca2b8e2c29fdf2a90e01fc668cf2ca6a4