Enable python3 support in devstack
We need to be able to choose between version v2 and v3 of python when we
run devstack.
This patchs adds a new variable to the gate-hook.sh script that permits
to pass a parameter to select the python version used in devstack gates.

gate_hook.sh <midonet-plugin> <python version>

For backwards compatibility with previous jenkins jobs if the python
version is ommitted it uses python2, if it receives "-py35" it will
use this version.

Related-Bug: #1670255

Change-Id: If0a0ddffb1323e76ca9dc151d12f57e692d5dae5
Signed-off-by: Antonio Ojea <aojea@midokura.com>