releasenotes: Ignore older entries in Pike release notes
I somehow had a wrong assumption that reno only includes entries
added during the cycle.  However, actually it also includes ones
modified in the cycle as well, as shown in the "reno list" output. [2]
I guess it would be useful to add an option to reno itself,
but for now, explicitly exclude entries for older releases,
following the similar change in horizon. [1]

[1] Iafe8365dca2a6ef963d7c459dab486975da939bc
[2] http://paste.openstack.org/show/617859/

Partial-Bug: #1682776
Change-Id: I8667f6432b158fe6b51483e3ae62d9c1ecd128e7