Avoid tox_install
Install the dependencies explicitly in requirements instead of using
tox_install which does not work well with OpenStack CI.

Adapt tox.ini for this.

This commit includes the following unrelated change to pass
the gate.
----------------------------------------
gate_hook: Remove neutron repo from the list of tempest plugins

The code has been moved to neutron-tempest-plugins.

Closes-Bug: #1769904
Change-Id: I1213f4c9947248ae61b9954fec22f77fc6bb5d2d
----------------------------------------

Co-Authored-By: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Iff427bbf9b659076e3bfecdd5e0bff9553db9dec