Port rally plugins to use proper modules
Sometime ago, all OpenStack related stuff moved from the main repository
to the separate one - https://github.com/openstack/rally-openstack .
All further development and maintance will be done only in a new
repository. As for first major release, the structure is not changed, so
only imports should be changed.

This commit includes the following unrelated change to
pass the gate.
----------------------------------------
Add dummy docstrings to avoid Sphinx 1.7.x warnings

We need to workaround this locally because:
- It's unclear how long os-vif patch [1] will take
  to be merged.
- While Neutron patch [2] has been merged, docs build on
  gate is using released versions of dependencies.

The idea was taken from keystone. [3]

[1] https://review.openstack.org/#/c/577028/
[2] https://review.openstack.org/#/c/577029/
[3] https://review.openstack.org/#/c/578121/

Related-Bug: #1771092
Change-Id: I2d125ce2251473f896fc251b8fa51cfa29721850
----------------------------------------

Co-Authored-By: YAMAMOTO Takashi <yamamoto@midokura.com>
Depends-On: I3446ccd5a36cb686282e85a9ea210d1723da1fef
Change-Id: I796e247bcbafc9c5b6d14e30a52e936b16157b8d