fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Also, bump pylint version as the old version doesn't seem working
well with python3. And fix new warnings/errors in the code.

Change-Id: I66b667d3ccdd6d6fe2415302f58125d2e3b7eb7b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>
Co-Authored-By: YAMAMOTO Takashi <yamamoto@midokura.com>