Rename test-requirements to avoid being picked by devstack
- Our test-requirements.txt has other neutron subprojects like
  neutron-fwaas for unit tests.
- Our devstack plugin uses pip_install to install ourselves.
- pip_install unconditionally installs requirements from
  test-requirements.

As a consequence, during a devstack run we installs neutron-fwaas
etc unconditionally. It isn't desirable because:

- We don't want to install neutron-fwaas etc for our non -full jobs.
  (Consider DB migration scripts, tempest plugins, etc.)
- The version in test-requirements.txt might be inappropriate for
  what we want to install for devstack based jobs.

To workaround the issue, this commit renames test-requirements.txt
to avoid being picked by pip_install.

tox-test-requirements.txt and doc/requirements.txt are identical
(actually a symlink) for now. We can separate and slim them down
a bit later if desirable.

An alternative is to stop using pip_install. But it might be more
complex than this commit given the amount of code it has.

Closes-Bug: #1814615
Change-Id: I52de37ee7a00e6a808cf1e33b4fdb1b3e462fa0b