Remove CINDER_PERIODIC_INTERVAL setting
Devstack allows setting the Cinder periodic_interval config option
via the CINDER_PERIODIC_INTERVAL shell var.  The dependency of this
patch removes the ability to set this via devstack, which means that
the Cinder periodic_interval will have its default value of 60.

The CINDER_PERIODIC_INTERVAL was introduced in devstack in havana as
a workaround for Bug #1180976.  That bug, however, was fixed in such
a way that it does not require adjusting the periodic_interval in
Cinder.

Depends-on: https://review.opendev.org/#/c/652711
Change-Id: I00c4a8aa0cca44b5940e011a61a0d51b46492c97
Partial-bug: #1824837