Join on explcit relationship paths
The join() in get_router_for_floatingip() is joining from entity
to entity without an explicit ON clause which creates an ambiguous
situation.  SQLAlchemy 1.3 guards against this now, so use the
real relationship-bound path so that the ORM does not need to guess.

Copied from https://review.opendev.org/#/c/642117/

Related-bug: #1819260
Change-Id: I9461ca708f05d037fb771b89ec7a95c403c5d1f6