Remove ocata jobs from branch
This repo configures two ocata jobs to run as periodic jobs, these
need to be only in the stable/ocata and not in this branch - remove
them.

The stable/ocata branch has these jobs configured, so the jobs continue
to run - but the Zuul setup is simplified.

Change-Id: I57958f8388505aee5654f20c0c9c60f0e3ceea15
(cherry picked from commit 431ad460f85136fcad773818f3b12f3a17917b95)