[ussuri][goal] Drop python 2.7 support and testing
OpenStack is dropping the Python 2.7 support in Ussuri cycle.

The networking-midonet project is ready with Python 3 and therefore it's
okay to drop Python 2.7 support. The only significant change we need to
make is to drop the centos7-based jobs since centos7 doesn't support
Python 3. These were non-voting and have been for some time, so this is
likely the best thing to do right now.

The complete discussion nad schedule can be found in

- http://lists.openstack.org/pipermail/openstack-discuss/2019-October/010142.html
- https://etherpad.openstack.org/p/drop-python2-support

The Ussuri community-wide goal can be found here:

https://governance.openstack.org/tc/goals/selected/ussuri/drop-py27.html

Change-Id: Ie95a528f4431d51cc551380d4d7066e4b6f299fa
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Co-authored-by: Ghanshyam Mann <gmann@ghanshyammann.com>